+++ public/d3_data_pie_ext.php
@@ -1,6 +1,6 @@
 <?php
 /*
-Copyright (C) Chris Park 2017
+Copyright (C) Chris Park 2017-2018
 diskover is released under the Apache 2.0 license. See
 LICENSE for the full license text.
  */
@@ -17,16 +17,10 @@ $mtime = getmtime($mtime);
 // get dir total size and file count
 $dirinfo = get_dir_info($client, $esIndex, $path, $filter, $mtime);
 
-// check for error
-if ($dirinfo[0] === 0) {
-    echo "Error: diskover index " . $esIndex . " has no data";
-    exit;
-}
-
 $data = [
             "name" => $path,
             "size" => $dirinfo[0],
-            "count" => $dirinfo[1],
+            "count" => $dirinfo[2],
             "children" => get_file_ext($client, $esIndex, $path, $filter, $mtime)
           ];