+++ public/hardlinks.php
@@ -31,13 +31,13 @@ require "d3_inc.php";
 	<body>
 		<?php include "nav.php"; ?>
 		<div class="container" id="error" style="display:none; margin-top:70px;">
-			<div class="row">
-				<div class="alert alert-dismissible alert-info col-xs-8">
-					<button type="button" class="close" data-dismiss="alert">&times;</button>
-					<i class="glyphicon glyphicon-exclamation-sign"></i> <strong>Sorry, no hardlinks found.</strong> Try searching for a different directory and clicking it's hardlinks analytics button, or changing filters (on filetree page) or minhardlinks value.
-				</div>
-			</div>
-		</div>
+            <div class="row">
+                <div class="alert alert-dismissible alert-info col-xs-8">
+                    <button type="button" class="close" data-dismiss="alert">&times;</button>
+                    <i class="glyphicon glyphicon-exclamation-sign"></i> <strong>Sorry, no hardlinks found.</strong> Try searching for a different directory and clicking it's hardlinks analytics button, or changing filters (on filetree page) or minhardlinks value.
+                </div>
+            </div>
+        </div>
 		<div class="container-fluid" id="mainwindow" style="margin-top: 70px;">
             <div id="hardlinkscharts-wrapper" style="display:none;">
                 <div class="row">